+++ cAER/trunk/events/imu9.h
@@ -124,7 +124,7 @@ static inline void caerIMU9EventValidate(caerIMU9Event event, caerIMU9EventPacke
 
 static inline void caerIMU9EventInvalidate(caerIMU9Event event, caerIMU9EventPacket packet) {
 	if (caerIMU9EventIsValid(event)) {
-		event->info &= htole16(~(U16T(1) << VALID_MARK_SHIFT));
+		event->info &= htole16((uint16_t)(~(U16T(1) << VALID_MARK_SHIFT)));
 
 		// Also decrease number of valid events. Number of total events doesn't change.
 		// Only call this on valid events!
